change plotanalyze dir name to plot_analyze and refactor all
